+++ asset.go
@@ -11,10 +11,10 @@ const assetsBasePath = "admin/themes"
 // of the Shopify API.
 // See: https://help.shopify.com/api/reference/asset
 type AssetService interface {
-	List(int, interface{}) ([]Asset, error)
-	Get(int, string) (*Asset, error)
-	Update(int, Asset) (*Asset, error)
-	Delete(int, string) error
+	List(int64, interface{}) ([]Asset, error)
+	Get(int64, string) (*Asset, error)
+	Update(int64, Asset) (*Asset, error)
+	Delete(int64, string) error
 }
 
 // AssetServiceOp handles communication with the asset related methods of
@@ -32,7 +32,7 @@ type Asset struct {
 	Size        int        `json:"size"`
 	SourceKey   string     `json:"source_key"`
 	Src         string     `json:"src"`
-	ThemeID     int        `json:"theme_id"`
+	ThemeID     int64      `json:"theme_id"`
 	Value       string     `json:"value"`
 	CreatedAt   *time.Time `json:"created_at"`
 	UpdatedAt   *time.Time `json:"updated_at"`
@@ -50,11 +50,11 @@ type AssetsResource struct {
 
 type assetGetOptions struct {
 	Key     string `url:"asset[key]"`
-	ThemeID int    `url:"theme_id"`
+	ThemeID int64  `url:"theme_id"`
 }
 
 // List the metadata for all assets in the given theme
-func (s *AssetServiceOp) List(themeID int, options interface{}) ([]Asset, error) {
+func (s *AssetServiceOp) List(themeID int64, options interface{}) ([]Asset, error) {
 	path := fmt.Sprintf("%s/%d/assets.json", assetsBasePath, themeID)
 	resource := new(AssetsResource)
 	err := s.client.Get(path, resource, options)
@@ -62,7 +62,7 @@ func (s *AssetServiceOp) List(themeID int, options interface{}) ([]Asset, error)
 }
 
 // Get an asset by key from the given theme
-func (s *AssetServiceOp) Get(themeID int, key string) (*Asset, error) {
+func (s *AssetServiceOp) Get(themeID int64, key string) (*Asset, error) {
 	path := fmt.Sprintf("%s/%d/assets.json", assetsBasePath, themeID)
 	options := assetGetOptions{
 		Key:     key,
@@ -74,7 +74,7 @@ func (s *AssetServiceOp) Get(themeID int, key string) (*Asset, error) {
 }
 
 // Update an asset
-func (s *AssetServiceOp) Update(themeID int, asset Asset) (*Asset, error) {
+func (s *AssetServiceOp) Update(themeID int64, asset Asset) (*Asset, error) {
 	path := fmt.Sprintf("%s/%d/assets.json", assetsBasePath, themeID)
 	wrappedData := AssetResource{Asset: &asset}
 	resource := new(AssetResource)
@@ -83,7 +83,7 @@ func (s *AssetServiceOp) Update(themeID int, asset Asset) (*Asset, error) {
 }
 
 // Delete an asset
-func (s *AssetServiceOp) Delete(themeID int, key string) error {
+func (s *AssetServiceOp) Delete(themeID int64, key string) error {
 	path := fmt.Sprintf("%s/%d/assets.json?asset[key]=%s", assetsBasePath, themeID, key)
 	return s.client.Delete(path)
 }
