+++ customcollection_test.go
@@ -16,7 +16,7 @@ func customCollectionTests(t *testing.T, collection CustomCollection) {
 		expected interface{}
 		actual   interface{}
 	}{
-		{"ID", 30497275952, collection.ID},
+		{"ID", int64(30497275952), collection.ID},
 		{"Handle", "macbooks", collection.Handle},
 		{"Title", "Macbooks", collection.Title},
 		{"BodyHTML", "Macbook Body", collection.BodyHTML},
